requirements.txt
django-widget-tweaks==1.4.8
Django
djangorestframework==3.11.0
django-import-export
Pillow==7.0.0
django-allauth

pytz==2019.3
requests==2.22.0

six==1.14.0

sqlparse==0.3.0
urllib3==1.25.8
Markdown==3.2.1
gunicorn==20.0.4
django-storages==1.8
boto==2.49.0
boto3==1.11.9
botocore==1.14.9
django-heroku==0.3.1
dj-database-url
whitenoise==5.0.1
psycopg2-binary
django-cors-headers==3.2.1